+++ packages/docs/docs/component/document/form/pro-el-form/demos/search-form.vue
@@ -9,7 +9,9 @@ import { type ProElFormColumn, ProElForm } from "@aiwen/ui"
 const columns: ProElFormColumn<any>[] = new Array(4).fill(1).map((_, i) => ({
   prop: "field" + i,
   label: "字段" + i,
-  colSpan: 12,
+  colProps: {
+    span:12
+  },
   fieldProps: {
     placeholder: "请输入字段" + i,
   },